requirements.txt
# Requirements для Ultra Code Manager Enhanced - Embedding System
# Все зависимости с фиксированными версиями для стабильности

# Основные библиотеки для эмбеддингов
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.0.1
transformers==4.33.2

# Дополнительные утилиты
scikit-learn==1.3.0
pandas==2.0.3
tqdm==4.66.1

# JSON обработка (если понадобится дополнительно)
ujson==5.8.0

# Логирование и мониторинг
psutil==5.9.5

# Опциональные зависимости для расширенного функционала
# Для работы с различными форматами файлов
chardet==5.1.0
python-magic==0.4.27

# Для веб-интерфейса (опционально)
flask==2.3.3
flask-cors==4.0.0

# Для визуализации эмбеддингов (опционально)
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Ускорение вычислений
numba==0.57.1

# CUDA поддержка (опционально, раскомментировать при наличии GPU)
# torch==2.0.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# faiss-gpu==1.7.4 